+++ lib/modules/auth/auth.module.dart
@@ -102,6 +102,10 @@ class AuthModule {
 
     AuthToken token = AuthToken.fromJson(tokenObject);
 
+    DateTime now = DateTime.now();
+
+    DateTime tokenCreatedAt = now.subtract(Duration(minutes: 2));
+
     userObject['token'] = token.accessToken;
     userObject['tokenType'] = token.tokenType;
     userObject['tokenExpiry'] = token.expiresIn;
@@ -110,8 +114,9 @@ class AuthModule {
     userObject['dirty'] = true;
     userObject['baseUrl'] = instanceUrl;
     userObject['authType'] = "token";
+    userObject['tokenCreatedAt'] = tokenCreatedAt.toIso8601String();
 
-    final user = User.fromApi(userObject);
+    User user = User.fromApi(userObject);
 
     await d2Instance.userModule.user.setData(user).save();