+++ lib/shared/utilities/http-details.util.dart
@@ -1,8 +1,52 @@
 import 'dart:convert';
 
 import 'package:d2_touch/modules/auth/entities/user.entity.dart';
+import 'package:d2_touch/modules/auth/models/auth-token.model.dart';
 import 'package:d2_touch/modules/auth/queries/user.query.dart';
+import 'package:dio/dio.dart';
 import 'package:sqflite/sqflite.dart';
+
+class OAuthToken {
+  String token;
+  int expiresIn;
+  DateTime tokenCreatedAt;
+  String refreshToken;
+  String url;
+
+  OAuthToken(
+      {required this.token,
+      required this.expiresIn,
+      required this.tokenCreatedAt,
+      required this.url,
+      required this.refreshToken});
+
+  bool isTokenExpired() {
+    final expirationTime = tokenCreatedAt.add(Duration(seconds: expiresIn));
+    return DateTime.now().isAfter(expirationTime);
+  }
+
+  Future<dynamic> renew() async {
+    BaseOptions options = BaseOptions(
+        baseUrl: url,
+        validateStatus: (status) => status == 200,
+        receiveDataWhenStatusError: true,
+        connectTimeout: const Duration(seconds: 60000),
+        receiveTimeout: const Duration(seconds: 60000));
+
+    try {
+      final dio = Dio(options);
+
+      final response = await dio.get(
+        '/oauth/token/refresh?token=${this.refreshToken}',
+      );
+      return response.data;
+    } on DioException catch (e) {
+      print(e.toString());
+      print(e.response?.data);
+      return e.response?.data ?? {};
+    }
+  }
+}
 
 class HttpDetails {
   Database? database;
@@ -35,6 +79,38 @@ class HttpDetails {
     this.token = user?.token;
     this.tokenType = user?.tokenType;
 
+//!TODO: Improve token refresh implementation not to use the SDK core
+//!TODO!: Improve token refresh implementation not to use the SDK core
+    if (this.token != null && this.tokenType == 'bearer') {
+      final OAuthToken token = OAuthToken(
+        token: this.token as String,
+        expiresIn: user?.tokenExpiry as int,
+        refreshToken: user?.refreshToken as String,
+        url: user?.baseUrl as String,
+        tokenCreatedAt: DateTime.parse(user?.tokenCreatedAt as String),
+      );
+
+      if (token.isTokenExpired() && user != null) {
+        dynamic tokenObject = await token.renew();
+        if (tokenObject != null && tokenObject['access_token'] != null) {
+          dynamic userObject = user.toJson();
+          AuthToken authToken = AuthToken.fromJson(tokenObject);
+          DateTime now = DateTime.now();
+          DateTime tokenCreatedAt = now.subtract(Duration(minutes: 2));
+          userObject['token'] = authToken.accessToken;
+          userObject['tokenType'] = authToken.tokenType;
+          userObject['tokenExpiry'] = token.expiresIn;
+          userObject['refreshToken'] = token.refreshToken;
+          userObject['isLoggedIn'] = true;
+          userObject['dirty'] = true;
+          userObject['authType'] = "token";
+          userObject['tokenCreatedAt'] = tokenCreatedAt.toIso8601String();
+          User userData = User.fromApi(userObject);
+          await UserQuery(database: database).setData(userData).save();
+          this.token = authToken.accessToken;
+        }
+      }
+    }
     return Future.value(this);
   }